// src/common/components/LoadingScreen/components/LoadingScreenSurface/LoadingScreenSurface.types.ts
/**
 * LoadingScreenSurface Props
 */
export type LoadingScreenSurfaceProps = {};

/**
 * LoadingScreenSurface State
 */
export type LoadingScreenSurfaceState = {};

/**
 * LoadingScreenSurface Styles
 */
export type LoadingScreenSurfaceStyles = {
    root: string;
    backdrop: string;
    logo: string;
    progressBar: string;
};